+++ app/javascript/vue/tasks/digitize/components/shared/depictions.vue
@@ -107,9 +107,11 @@ export default {
       if (newVal.id && newVal.id != oldVal.id) {
         this.$refs.depiction.setOption('autoProcessQueue', true)
         this.$refs.depiction.processQueue()
-        this.getDepictions(newVal.id).then((response) => {
-          this.figuresList = response.body
-        })
+        this.getDepictions(newVal.id)
+          .then((response) => {
+            this.figuresList = response.body
+          })
+          .catch(() => {})
       } else if (!newVal.id) {
         this.figuresList = []
         this.$refs.depiction.setOption('autoProcessQueue', false)