+++ app/javascript/vue/tasks/digitize/store/actions/getCollectionObject.js
@@ -2,11 +2,14 @@ import { CollectionObject } from '@/routes/endpoints'
 import { MutationNames } from '../mutations/mutations'
 import SetParam from '@/helpers/setParam'
 
-export default ({ commit }, id) =>
-  new Promise((resolve, reject) => {
-    SetParam('/tasks/accessions/comprehensive', 'collection_object_id', id)
-    CollectionObject.find(id).then((response) => {
-      commit(MutationNames.SetCollectionObject, response.body)
-      resolve(response.body)
-    })
+export default ({ commit }, id) => {
+  SetParam('/tasks/accessions/comprehensive', 'collection_object_id', id)
+
+  const request = CollectionObject.find(id)
+
+  request.then(({ body }) => {
+    commit(MutationNames.SetCollectionObject, body)
   })
+
+  return request
+}