+++ CHANGELOG.md
@@ -15,10 +15,12 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 - Contributing link [#3752]
 - Uncaught promise errors [#3767]
+- Custom attributes triggers `isUpdated` ce state [#3764]
 
 [#3752]: https://github.com/SpeciesFileGroup/taxonworks/issues/3752
 [#2718]: https://github.com/SpeciesFileGroup/taxonworks/issues/2718
 [#3767]: https://github.com/SpeciesFileGroup/taxonworks/issues/3767
+[#3764]: https://github.com/SpeciesFileGroup/taxonworks/issues/3764
 
 ## [0.37.1] - 2024-01-04